Authentication App contains 1. Sign in 2. Sign out 3. Firebase 4. Redux store to save users
