+++ src/components/Header.jsx
@@ -1,14 +1,16 @@
-import React from 'react';
-import { signOut } from "firebase/auth";
+import React,{useEffect} from 'react';
+import { signOut, onAuthStateChanged } from "firebase/auth";
 import { auth } from '../utils/firebase';
 import { useNavigate } from 'react-router-dom';
-import { useSelector } from 'react-redux';
-
+import { useDispatch, useSelector } from 'react-redux';
+import { addUser, removeUser } from '../utils/userSlice';
+import { NETFLIX_LOGO } from '../utils/constant';
 
 const Header = () => {
   const navigate = useNavigate();
   const user = useSelector(store => store.user);
-
+  const dispatch = useDispatch();
+  
   const handleSignOut = () => {
     // Implement sign out logic here
     signOut(auth).then(() => {
@@ -18,11 +20,36 @@ const Header = () => {
       navigate("/error");
     });
   }
+
+  useEffect(() =>{
+    const unsubscribe = onAuthStateChanged(auth, (user) => {
+      if (user) {
+        // User is signed in
+        const {uid, email, displayName, photoURL} = user;
+        dispatch(addUser({
+          udi: uid, 
+          email: email, 
+          displayName: displayName, 
+          photoURL: photoURL
+        })
+      );
+      navigate("/browse");
+      } else {
+        // User is signed out
+        dispatch(removeUser());
+        navigate("/");
+      }
+
+      //Unsubscribe when component unmount
+      return () => unsubscribe();
+    });
+    
+  },[])
   return (
-    <>
+    <>  
       <div className='absolute px-8 py-2 z-10 w-screen flex justify-between bg-gradient-to-b from-black'>
         <img
-          src="https://cdn.cookielaw.org/logos/dd6b162f-1a32-456a-9cfe-897231c7763c/4345ea78-053c-46d2-b11e-09adaef973dc/Netflix_Logo_PMS.png"
+          src={NETFLIX_LOGO}
           alt='Netflix'
           className='max-w-60' />
         {user && <div className='flex p-2'>
